// src/validators/auth.schema.ts
import * as yup from 'yup';

export const registerSchema = yup.object().shape({
  email: yup.string().email('Invalid email address').required('Email is required'),
  password: yup.string().min(6, 'Password must be at least 6 characters').required('Password is required'),
  role: yup.string().oneOf(['CLIENT', 'ADMIN'], 'Invalid role').default('CLIENT'),
});

export const loginSchema = yup.object().shape({
  email: yup.string().email('Invalid email address').required('Email is required'),
  password: yup.string().required('Password is required'),
});

export type RegisterData = yup.InferType<typeof registerSchema>;
export type LoginData = yup.InferType<typeof loginSchema>;